refactor(mcp-server): a tool is a file, the way a command is
The layout the cli already has: one file per thing the surface offers, under a
tree that reads as the thing it offers, and what more than one of them needs in
`lib/` beside them.

`createNibrunMcpServer` stops being exported. The handler is its only caller,
and a factory nothing outside builds is not an entry point.
