+++ packages/mcp-server/src/index.ts
@@ -1,4 +1,3 @@
 /** biome-ignore-all lint/performance/noBarrelFile: index is the only allowed file where we can export other files */
 
 export { type ApiForCaller, createNibrunMcpHandler } from '#handler.ts';
-export { createNibrunMcpServer } from '#server.ts';
